--- packages/editor/src/math/index.ts
@@ -1,1 +1,0 @@
-export * from './editor'
